refactor rdf into module
